libs/gl-testing/pyproject.toml
[tool.poetry]
name = "gltesting"
version = "0.1.0"
description = ""
authors = ["Christian Decker <decker@blockstream.com>"]
license = "MIT"

packages = [
    { include = "gltesting" },
]

[tool.poetry.dependencies]
python = "^3.8"
cryptography = ">=36.0.1,<37.0.0"
ephemeral-port-reserve = "^1.1.4"
sh = "^1.14.2"
pytest-timeout = "^2.1.0"
# `pip` doesn't like relative-path dependencies. Should be ok once we
# publish this on PyPI so pip can find it too.
# gl-client-py = {path = "../gl-client-py"}
pytest-xdist = "^2.5.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
rich = "^12.5.1"
pyln-testing = "^23.11"
pyln-grpc-proto = "^0.1"
protobuf = ">=3"
flaky = "^3"
anyio = "^3.7.1"
purerpc = { version = "^0.8.0", extras = ["grpc"] }
pytest-sugar = "^0.9.7"

[tool.poetry.dev-dependencies]
mypy = "^1"
typed-ast = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
